requirements.txt
pyyaml==6.0.1
feedparser==6.0.10
requests==2.31.0
schedule==1.2.0 